refactor(cli): collapse list to a concrete oclif command (#2197)
## Summary
This stacked PR now builds on #2189 by collapsing `nemoclaw list` down
to a single concrete oclif command and a small generic oclif runner
helper. Because #2192 merged independently before this stack was fully
rebased, the branch was refreshed against the latest #2189 head to keep
the stack mergeable while preserving the same final `list` behavior.

## Changes
- Replace the generated `createListCommand(...)` path in
`src/lib/list-command.ts` with a single exported `ListCommand` class
that builds its own runtime dependencies via `buildListCommandDeps()`.
- Add `src/lib/oclif-runner.ts` so the legacy dispatcher can run
registered oclif commands without keeping list-specific wrapper logic.
- Simplify `src/lib/list-command-deps.ts` to return only the inventory
dependencies actually needed by `getSandboxInventory()`.
- Update `src/lib/oclif-commands.ts` and `src/nemoclaw.ts` to use the
concrete command class and the generic oclif runner.
- Remove `src/lib/list-command.test.ts`; the remaining inventory unit
tests plus CLI integration tests cover the list behavior directly.
- Refresh the branch against the latest #2189 head after #2192 landed
separately, so GitHub can merge the stack cleanly.

## Type of Change
- [x] Code change (feature, bug fix, or refactor)
- [ ] Code change with doc updates
- [ ] Doc only (prose changes, no code sample modifications)
- [ ] Doc only (includes code sample changes)

## Verification
- [x] `npx prek run --all-files` passes
- [x] `npm test` passes
- [x] Tests added or updated for new or changed behavior
- [x] No secrets, API keys, or credentials committed
- [ ] Docs updated for user-facing behavior changes
- [ ] `make docs` builds without warnings (doc changes only)
- [ ] Doc pages follow the [style
guide](https://github.com/NVIDIA/NemoClaw/blob/main/docs/CONTRIBUTING.md)
(doc changes only)
- [ ] New doc pages include SPDX header and frontmatter (new pages only)

## AI Disclosure
- [x] AI-assisted — tool: OpenAI Codex

---

---------
