+++ src/lib/oclif-commands.ts
@@ -1,11 +1,8 @@
 // SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { createListCommand, type ListCommandClass } from "./list-command";
-import { buildListCommandDeps } from "./list-command-deps";
+import ListCommand from "./list-command";
 
-const commands: Record<string, ListCommandClass> = {
-  list: createListCommand(buildListCommandDeps),
+export default {
+  list: ListCommand,
 };
-
-export default commands;
